+++ demo/demo.js
@@ -67,13 +67,14 @@ $( function () {
 
 	JS.Responsive.setLocation( 48.136609, 17.107228 );
 
-	JS.Responsive.init({
-		detectAdblock: {
-			adblockDetectedClass: 'blocker-detected',
-			noAdblockClass: 'no-blocker',
-		},
-		all: true
-	});
+	JS.Responsive.init();
+	// JS.Responsive.init({
+	// 	detectAdblock: {
+	// 		adblockDetectedClass: 'blocker-detected',
+	// 		noAdblockClass: 'no-blocker',
+	// 	},
+	// 	all: true
+	// });
 
 } );